Fixes and cleanups
